Use `Create` with `AutoInit` before `Reconcile`
go-git-providers has at present some issues with the idempotency of the
available `Reconcile` methods, and setting e.g. the default branch
correctly.

Using `Create` with `AutoInit` works around this, with as
downside that a `main` branch will always be created, even if the user
has configured a different target bootstrap branch.

Signed-off-by: Hidde Beydals <hello@hidde.co>